% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip.data.frame.R
\name{flip.data.frame}
\alias{flip.data.frame}
\title{Check and flip the alleles of a data frame}
\usage{
flip.data.frame(df)
}
\arguments{
\item{df}{data frame. With 5 columns in order: rsID, a1, a0 (genotype to be checked), a1, a0 (genotype to be checked against)}
}
\description{
Lower level function which takes an input data-frame with alleles from two separate genotypes
and then compares the alleles and flips them if necessary.
}
\details{
flip.data.frame is called within other major functions and is required for aligning the direction of
effects between two different datasets for performing a regression between them. The function (and the
other functions it calls) operates in two ways: first checking whether alleles a1/a0 in one dataset match
a1/a0 in the other. And then it switches the strand of the alleles being checked to make sure that these do
not mismatch with those on the other side either.
}
