% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpaths_frmEnds.R
\name{getpaths_frmEnds}
\alias{getpaths_frmEnds}
\title{Get (KEGG) biological pathways from end point}
\usage{
getpaths_frmEnds(gene_entrez)
}
\arguments{
\item{gene}{character or numeric (Entrez ID)}
}
\description{
getpaths_frmEnds returns a list of (KEGG) pathways from an input gene end-point.
}
\details{
This function requires the input of a single gene (in Entrez ID format) from this it will
search the KEGG database for any pathways which include the gene. This is done by converting the
Entrez ID into a KEGG human gene ID (E.g. 1234 becomes hsa:1234).
The pathways in which the gene is present are then filtered to only include those where the gene
is an "end-point" for the pathway: I.e. it has one or more inward-facing edges and no outward-facing
edges.
These pathways are those which are passed to return()
If no pathways are found or the gene is present in pathways but is not an end-point, the function
will inform the user and will return nothing.
}
\examples{
## Search for pathways in KEGG which have IL18 as an end-point.
getpaths_frmEnds(3606)

}
