% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_enet_mr.R
\name{glmnet_enet_mr}
\alias{glmnet_enet_mr}
\title{Run elastic net-based mendelian randomisation from input MR object}
\usage{
glmnet_enet_mr(
  object,
  cv = TRUE,
  lambda = 0.1,
  alpha = 0.1,
  cv.param = "lambda.1se"
)
}
\arguments{
\item{object}{MR Input object. Contains a matrix of exposure SNPs to betas, exposure SNPs to SEs and then a list of outcome SNP betas and SEs.}
}
\description{
Lower level function. Takes input "MR input" object (created by mr_mvinput()) and performs elastic net mendelian
randomisation with this input. Outputs a model of exposures and betas.
}
\details{
This function was written by Dr. Verena Zuber, <v.zuber(at)imperial.ac.uk> who has kindly given us permission
to incorporate it into the PathWAS package.
The original code for this can be found here: https://github.com/verena-zuber/demo_AMD/blob/master/mvMR_glmnet.R
It works in a similar method to the MendelianRandomization R package in working from the same input object
(a matrix of betas and SEs for every SNP used with each exposure).
}
