% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heading.R
\name{heading}
\alias{heading}
\title{Print as heading}
\usage{
heading(sentence)
}
\arguments{
\item{sentence}{character. A string of variable length for printing.}
}
\description{
heading prints an input string in a "heading" format
}
\details{
A function primarily for aesthetics within the PathWAS pipeline and for progress checks throughout.
}
