% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_geneglob.R
\name{kegg_geneglob}
\alias{kegg_geneglob}
\title{Obtain all genes from a KEGG pathway}
\usage{
kegg_geneglob(pathway, in_mart = NULL)
}
\arguments{
\item{pathway}{character. KEGG pathway ID.}

\item{in_mart}{data frame. A data frame of multiple columns which, at a minimum includes entrezgene_id and external_gene_name.}
}
\description{
Returns all genes associated with input pathway (excluding directionality)
}
\details{
This function takes a pathway (possibly defined by the getpaths_frmEnds function) from KEGG and then finds
all genes associated with this pathway and outputs that.
Depending on whether or not you have an available BioMart reference table, this can be either as a list of Entrez IDs or as a data frame
of KEGG IDs (which is simply "hsa:ENTREZ-ID") and gene names.
A BioMart table can be obtained either by downloading a full dataset from https://www.ensembl.org/biomart/martview/ or through usage of
the biomaRt package. This table MUST have the entrezgene_id and external_gene_name columns (default column names in BioMart for Entrez
and gene names).
}
