% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmp.R
\name{lmp}
\alias{lmp}
\title{Extract P-value from regression model}
\usage{
lmp(modelobject)
}
\arguments{
\item{modelobject}{model output from lm function}
}
\description{
lmp returns the P-value of an input model
}
\details{
A lower level function for outputting the P-value for a model, as this is not always immediately
extractable from just the one of the variables of the model.
}
