% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_MungeR.R
\name{omics_MungeR}
\alias{omics_MungeR}
\title{Merging clumped SNPs with the SNPs from an omics end-point}
\usage{
omics_MungeR(
  endpoint_omics,
  clumped_snplist,
  omics_SNPCol = "rsid",
  omics_EffAllCol = "a1",
  omics_OthAllCol = "a0",
  clumped_SNPCol = "rsid",
  clumped_EffAllCol = "a1",
  clumped_OthAllCol = "a0"
)
}
\arguments{
\item{endpoint_omics}{Data frame of summary stats of your selected omics. Contains, at minimum a SNP ID column, effect allele and reference allele.}

\item{clumped_snplist}{Data frame of summary stats of clumped SNPs. Contains, at minimum a SNP ID column, effect allele and reference allele.}

\item{omics_SNPCol}{Character. Name/number of column containing the SNP IDs in the omics data frame.}

\item{omics_EffAllCol}{Character. Name/number of column containing the effect allele in the omics data frame.}

\item{omics_OthAllCol}{Character. Name/number of column containing the alternative allele in the omics data frame.}

\item{clumped_SNPCol}{Character. Name/number of column containing the SNP IDs in the clumped SNPs data frame.}

\item{clumped_EffAllCol}{Character. Name/number of column containing the effect allele in the clumped SNPs data frame.}

\item{clumped_OthAllCol}{Character. Name/number of column containing the alternative allele in the clumped SNPs data frame.}
}
\description{
omics_MungeR inputs a data frame of clumped SNPs from qtl_clumpR and the summary statistics from your selected
end-point omics and then determines the alignment.
}
\details{
This is an intermediate step of PathWAS, required for setting up the next stages. The function inputs both data frames of the summary stats for your chosen end-point omics and the clumped SNPs,
most likely your output from qtl_clumpR. It combines these two data frames and produces a merged data frame
of the omics summary stats with an additional column which will be used to determine whether the SNP effects need to be flipped in order
to align the effects with those from your clumped SNPs.
At a minimum the input data frames must have a SNP ID column (usually rsID) and an effect allele and alternative
allele column. These can be defined manually as part of the function but default to "rsid,a1,a0" respectively.
}
\examples{
## Input two different data frames with different column names for merging
omics_MungeR(endpoint_omics = my_proteomics_stats, clumped_snplist = my_snps_dataframe, omics_SNPCol = "SNPID",
omics_EffAllCol = "Effect_Allele", omics_OthAllCol = "Other_Allele", clumped_SNPCol = "rsid", clumped_EffAllCol = "a1",
clumped_OthAllCol = "a0")

}
