% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smple_paths.R
\name{smple_paths}
\alias{smple_paths}
\title{Get all simple paths from KEGG pathway}
\usage{
smple_paths(
  pathway,
  gene_entrez,
  keep_routes = TRUE,
  hsapien_mart,
  saveDir = "./",
  delete_tmps = FALSE,
  cut_paths = -1
)
}
\arguments{
\item{pathway}{character. KEGG pathway ID (e.g.: path:hsa05131).}

\item{gene_entrez}{character or numeric. Entrez gene ID.}

\item{keep_routes}{logical. Should the function output all simple pathways, or output a list of genes.}

\item{hsapien_mart}{data frame. Data frame of gene names including at least entrezgene_id, external_gene_name.}

\item{saveDir}{character. Directory for saving the KGML file downloaded from KEGG}

\item{delete_tmps}{logical. After running, delete the KGML downloaded or not.}

\item{cut_paths}{numeric. Cut-off for length of simple paths (default is -1 for all_smple_paths). Add threshold to reduce computing time.}
}
\description{
smple_paths returns a list of all simple (one-route) pathways through a larger/more complex bioogical pathway to a chosen end point.
}
\details{
This function takes a pathway (possibly defined by the getpaths_frmEnds function) from KEGG and then finds
all simple "routes" that lead towards a chosen end-point. Therefore it searches for every route through the pathway
and searches for every gene included in those routes that lead towards the end-point.
Therefore this outputs only the genes from the pathway which are associated with the end-point, and excludes genes which
theoretically have no influence over the end-point.
}
\examples{
## Find all simple routes in pathway hsa05131 (Shigellosis) leading to end point 3606 (IL18).
## Output only the genes involved in those routes (I.e. do not keep the routes themselves).

smple_paths(pathway = path:hsa05131, gene_entrez = 3606, keep_routes = FALSE)

}
