% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_check.R
\name{transcript_check}
\alias{transcript_check}
\title{Check whether a gene is transcribed via the pathway (I.e. via interaction with DNA)}
\usage{
transcript_check(pathway, gene_entrez, kegg_file = "./")
}
\arguments{
\item{pathway}{character. KEGG pathway ID (e.g.: hsa05131).}

\item{gene_entrez}{character or numeric. Entrez gene ID.}

\item{kegg_file}{character. Full file location of a KEGG KGML file.}
}
\description{
transcript_check returns a single row of a data frame, containing the interactions leading to it in KEGG.
}
\details{
This function uses a defined KEGG KGML, pathway name, and Entrz gene. It searches the KGML for this gene
(only in nodes which the gene is at the end of an interaction) and then returns those rows to the function
user. If the "type" of interaction is GErel, this is defined by KEGG as "gene expression interaction,
indicating relation of transcription factor and target gene product" and therefore, this is a gene product
which is transcribed through the pathway.
}
