% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smple_paths.R
\name{smple_paths}
\alias{smple_paths}
\title{Get all simple paths from KEGG pathway}
\usage{
smple_paths(pathway, gene_entrez, keep_routes = TRUE)
}
\arguments{
\item{pathway}{character. KEGG pathway ID.}

\item{gene_entrez}{character or numeric. Entrez gene ID.}

\item{keep_routes}{logical. Should the function output all simple pathways, or output a list of genes.}
}
\description{
smple_paths returns a list of all simple (one-route) pathways through a larger/more complex bioogical pathway to a chosen end point.
}
\details{
This function takes a pathway (possibly defined by the getpaths_frmEnds function) from KEGG and then finds
all simple "routes" that lead towards a chosen end-point. Therefore it searches for every route through the pathway
and searches for every gene included in those routes that lead towards the end-point.
Therefore this outputs only the genes from the pathway which are associated with the end-point, and excludes genes which
theoretically have no influence over the end-point.
}
\examples{
## Find all simple routes in pathway hsa05131 (Shigellosis) leading to end point 3606 (IL18).
## Output only the genes involved in those routes (I.e. do not keep the routes themselves).
smple_paths(pathway = hsa05131, gene_entrez = 3606, keep_routes = FALSE)
}
